Move CharacterPool to models/game/pools
