+++ frontend/models/game/GameBuilder.php
@@ -10,8 +10,8 @@ namespace frontend\models\game;
 
 
 use frontend\models\game\base\BeeTypesInterface;
-use frontend\models\game\base\CharacterPool;
 use frontend\models\game\base\HoneyPool;
+use frontend\models\game\pools\CharacterPool;
 use frontend\models\game\characters\Drone;
 use frontend\models\game\characters\Healer;
 use frontend\models\game\characters\Player;
