+++ tests/game/inside/CharacterPoolTest.php
@@ -10,7 +10,7 @@ namespace tests\game\inside;
 
 
 use frontend\models\game\base\BeeInterface;
-use frontend\models\game\base\CharacterPool;
+use frontend\models\game\base\CharacterPoolInterface;
 use frontend\models\game\characters\Drone;
 use frontend\models\game\characters\Healer;
 use frontend\models\game\characters\Player;
@@ -22,7 +22,7 @@ use tests\fixtures\GameWithoutCharacters;
 
 class CharacterPoolTest extends \PHPUnit_Framework_TestCase
 {
-    /** @var  CharacterPool */
+    /** @var  CharacterPoolInterface */
     private static $pool;
     /** @var  Game */
     private static $game;
